tests: add integration test for private link in sidebar
