+++ packages/web-app-files/src/components/SideBar/Links/FileLinks.vue
@@ -7,7 +7,7 @@
     >
       <oc-loader v-if="linksLoading" :aria-label="$gettext('Loading list of file links')" />
       <template v-else>
-        <private-link-item />
+        <private-link-item v-if="privateLinkEnabled" />
         <h4 v-translate class="oc-text-bold oc-m-rm oc-text-initial">Public Links</h4>
         <div v-if="canCreatePublicLinks" class="oc-my-s">
           <p v-translate class="oc-text-muted">
@@ -186,6 +186,10 @@ export default {
     },
     $_addButtonAriaLabel() {
       return this.$gettext('Create new public link')
+    },
+
+    privateLinkEnabled() {
+      return this.capabilities.files.privateLinks
     }
   },
   watch: {
